Updated admin authentication and added new admin creation features
Updated the admin authentication endpoint in `Authenticate.cs`. Added new classes `AdminCreatedHandler` in `RSAllies.Mail.Features` and `RSAllies.SMS.Features` for handling admin creation notifications via email and SMS. Introduced `AdminCreated` in `RSAllies.Shared.Notifications` to represent the notification sent when a new admin is created. Added `CreateAdminDto` in `RSAllies.Users.Contracts.Requests` as the data transfer object for creating a new admin. Created an abstract class `CreateAdmin` in `RSAllies.Users.Features.Admin` for admin creation command and handler, and to check for existing admin with the same username, phone, or email. Added `CreateAdminEndPoint` in `CreateAdmin.cs` for the endpoint `/api/admin` that accepts a POST request with a `CreateAdminDto` object, sends a command to create a new admin, and returns a success or error result.