Refactor session query by region to use Guid
Refactored the `GetSessionsByRegionAndDate` class to improve session querying by transitioning from using region names to `Guid` identifiers for regions. This includes changes in the `Query` class property from `Region` to `RegionId`, updating LINQ queries in the `Handler` class to filter sessions by `RegionId`, and modifying the endpoint definition in `GetSessionsByRegionAndDateEndPoint` to use `regionId` of type `guid` in the URL path. Fixed an incorrect initialization of `RegionId` to `string.Empty` which should be `Guid.Empty`.