Add semaphore to synchronize email sending
Implemented a SemaphoreSlim in the SessionJob class to ensure that email sending operations are executed one at a time in a concurrent environment. This addition involves using the semaphore to control access to the smtpClient.SendAsync method, ensuring that multiple tasks or threads do not send emails simultaneously. The semaphore is initialized with a count of 1 to allow only one thread to execute the critical section at a time and is properly released after an email is sent to allow the next operation to proceed.